Move the scoring out of the CLI, and score per placement
Review follow-ups. Two of these broke things I had claimed were verified.

The test used a top-level await import, which is illegal under this
repo's module: es2020 -- `tsc --build` fails with TS1378, so the CI build
job would have failed. I missed it because I ran build-incremental with
its output discarded and never checked the exit code, then ran jest
against the stale compiled output from before the change. That is the
same stale-artifact trap I had just written a commit note about, reached
from the other direction.

Importing model_report.mjs also ran the CLI inside the test runner: it
has no entry-point guard and executes `process.exitCode = await main()`
at module scope. So the suite dumped the usage header to stderr and jest
exited 1 on a leftover exit code even with every test green.

Both are fixed by extracting the pure scoring into
scripts/debug/displacement.mjs, which has no side effects on import. The
test imports that; model_report.mjs imports it too.

collectDisplacement no longer dedups by geometry localID. That dedup is
copied from collectMeshes, where it is free because local bounds are
identical for every placement -- but here the placement IS the
measurement. extractMappedItem and the AP214 occurrence stack emit one
walk entry per instance, so deduping hid every flung-away instance of a
shared window or fastener, and computed the robust centre from an
unrepresentative sample.

stage.fired now counts meshes WALKED rather than meshes placed. As
written it did the opposite of its own comment: a model whose every
placement was unusable reported "PROBE NEVER FIRED" and exited 2. It
also made fired identical to measured, hiding skipped meshes.

A transform of the right shape holding NaN now drops that mesh instead
of throwing -- that is data, not a programming error. A genuinely wrong
walk-tuple index still throws, and main() catches it per-stage so one
stage's failure cannot discard the curve/loop/face/mesh reports the same
run already collected.

Documents the one caveat of reusing the shared threshold: distances can
be exactly zero, so a model with over half its meshes at the robust
centre gives median 0 and flags everything. --limit is the escape.

Verified with exit codes actually checked this time: tsc 0, build 0,
jest 0 (6 tests), CLI 0, eslint clean on src and on both changed script
files, suite 77 suites / 382 tests.
